// models/devices-module/devices/types.ts
export interface DeviceInterface {
  id: string
  type: DeviceEntityTypes

  draft: boolean

  parentId: string | null

  identifier: string
  name: string | null
  comment: string | null

  state: DeviceStateTypes
  enabled: boolean

  control: Array<string>

  owner: string | null

  relationshipNames: Array<string>

  children: Array<DeviceInterface>
  channels: Array<ChannelInterface>
  properties: Array<DevicePropertyInterface>
  configuration: Array<DeviceConfigurationInterface>

  hardware: HardwareInterface | null
  firmware: FirmwareInterface | null

  isEnabled: boolean
  isReady: boolean
  icon: string
  title: string
  hasComment: boolean
}

// ENTITY TYPES
// ============

export enum DeviceEntityTypes {
  PHYSICAL = 'devices-module/physical-device',
}

export enum DeviceStateTypes {
  CONNECTED = 'connected',
  DISCONNECTED = 'disconnected',
  INIT = 'init',
  READY = 'ready',
  SLEEPING = 'sleeping',
  LOST = 'lost',
  ALERT = 'alert',
  UNKNOWN = 'unknown',
}